did some stuff

+++ run_svd.m
@@ -9,5 +9,5 @@ function svd_dat = run_svd(svd_dat,upper,lower)
     upper = ceil(upper*shape(3));
     S(1:upper,1:upper) = 0;
 
-    svd_dat = reshape(U*S*V', shape(1), shape(2), shape(3));
+    svd_dat = single(reshape(U*S*V', shape(1), shape(2), shape(3)));
 end
